refactor: move all test files to __tests__ directories
- Move 29 co-located .test.ts files into __tests__ directories
- Update all import paths in moved test files
- Add lint:test-location script to enforce test file placement
- Add test location check to CI (pr-checks.yml)
- Add test location check to pre-commit hook
- Update vitest.config.ts to exclude __tests__ directories from coverage
- Remove duplicate SymbolTable.test.ts (kept larger version)
- Remove duplicate TypeValidator.test.ts (kept larger version)
