+++ .husky/pre-commit
@@ -1,1 +1,2 @@
 npx lint-staged
+npm run lint:test-location
