+++ src/transpiler/output/codegen/generators/support/__tests__/CommentUtils.test.ts
@@ -3,11 +3,11 @@
  * Tests the comment handling utility functions
  */
 import { describe, it, expect, vi } from "vitest";
-import commentUtils from "./CommentUtils";
-import CommentExtractor from "../../../../logic/analysis/CommentExtractor";
-import CommentFormatter from "../../CommentFormatter";
-import ECommentType from "../../../../types/ECommentType";
-import IComment from "../../../../types/IComment";
+import commentUtils from "../CommentUtils";
+import CommentExtractor from "../../../../../logic/analysis/CommentExtractor";
+import CommentFormatter from "../../../CommentFormatter";
+import ECommentType from "../../../../../types/ECommentType";
+import IComment from "../../../../../types/IComment";
 
 const {
   getLeadingComments,
