+++ src/utils/__tests__/BitUtils.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import BitUtils from "./BitUtils";
+import BitUtils from "../BitUtils";
 
 // ========================================================================
 // boolToInt
