+++ src/utils/__tests__/LiteralUtils.test.ts
@@ -4,12 +4,12 @@
  */
 import { describe, it, expect } from "vitest";
 import { CharStream, CommonTokenStream } from "antlr4ng";
-import { CNextLexer } from "../transpiler/logic/parser/grammar/CNextLexer";
+import { CNextLexer } from "../../transpiler/logic/parser/grammar/CNextLexer";
 import {
   CNextParser,
   LiteralContext,
-} from "../transpiler/logic/parser/grammar/CNextParser";
-import LiteralUtils from "./LiteralUtils";
+} from "../../transpiler/logic/parser/grammar/CNextParser";
+import LiteralUtils from "../LiteralUtils";
 
 /**
  * Helper to parse C-Next code and extract the first literal from a variable declaration.
